requirements.txt
python-dotenv
wikitextparser
langchain
langchain-google-genai
langchain-text-splitters
pinecone-client
chainlit
langchain-pinecone